move get_base_campus function to the caller file
